+++ src/errors/index.ts
@@ -30,10 +30,8 @@ export class EmptySequentialValuesError extends ChacaError {
 }
 
 export class PickFieldDefinitionError extends ChacaError {
-  constructor(public readonly fieldRoute: string) {
-    super(
-      `On field '${fieldRoute}'. A pick field should must have a number of options to choose from greater than or equal to the 'count' property`,
-    );
+  constructor(readonly fieldRoute: string, message: string) {
+    super(`On field '${fieldRoute}'. ${message}`);
 
     this.name = "ChacaError.PickFieldDefinitionError";
   }